training/requirements.txt
torch>=2.0
transformers>=4.40
datasets>=2.18
optimum[onnxruntime]>=1.18
evaluate>=0.4
seqeval>=1.2
pyyaml>=6.0
accelerate>=0.28
